# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Library",
        "description": "Library nodeJS Application API",
        "license": "ISC"
    },
    "basePath": "/api/v1",
    "tags": [
        {
            "name": "Book",
            "description": "API for book in the system"
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/book/create": {
            "post": {
                "tags": [
                    "Book"
                ],
                "description": "Create new book in systems",
                "parameters": [
                    {
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "description": "Book that we want to create",
                        "example": {
                            "title": "cien años de soledad",
                            "pages": 402,
                            "year": 1962,
                            "isbn": "123-456-789-A1",
                            "author": "Gabriel Garcia Marquez"
                        },
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "New Book is created",
                        "example": {
                            "application/json": {
                                "status": true,
                                "message": "Book created!"
                            }
                        }
                    }
                }
            }
        },
        "/book/list": {
            "get": {
                "tags": [
                    "Book"
                ],
                "responses": {
                    "200": {
                        "description": "all books are returned",
                        "examples": {
                            "application/json": {
                                "status": true,
                                "listBooks": "array"
                            }
                        }
                    }
                }
            }
        },
        "/book/{isbn}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "responses": {
                    "200": {
                        "description": "A book returned",
                        "examples": {
                            "application/json": {
                                "status": true,
                                "bookByISBN": "object"
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "isbn",
                    "in": "path",
                    "required": true,
                    "description": "book isbn to find",
                    "type": "string"
                }
            ]
        },
        "/book/delete/{isbn}": {
            "delete": {
                "tags": [
                    "Book"
                ],
                "responses": {
                    "200": {
                        "description": "A book deleted!",
                        "examples": {
                            "application/json": {
                                "status": true,
                                "message": "Book deleted!"
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "isbn",
                    "in": "path",
                    "required": true,
                    "description": "book isbn to delete",
                    "type": "string"
                }
            ]
        },
        "/book/update/{isbn}": {
            "put": {
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "description": "Book to update",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A book update",
                        "examples": {
                            "application/json": {
                                "status": true,
                                "message": "Book Updated!"
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "isbn",
                    "in": "path",
                    "required": true,
                    "description": "book isbn to update",
                    "type": "string"
                }
            ]
        }
    },
    "definitions": {
        "Book": {
            "required": [
                "title",
                "pages",
                "year",
                "isbn",
                "author"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "uniqueItems": false,
                    "example": "Cien años de soleda"
                },
                "pages": {
                    "type": "int",
                    "example": 402
                },
                "year": {
                    "type": "int",
                    "example": 1962
                },
                "isbn": {
                    "type": "string",
                    "uniqueItems": false,
                    "example": "123-456-789-A1"
                },
                "author": {
                    "type": "string",
                    "uniqueItems": false,
                    "example": "Gabriel Garcia Marquez"
                }
            }
        }
    }
}